clean up: create helper_classes.py, merge dynamic_biobuilder and dynamic_biobuilder_from_timeline
