+++ bw_timex/__init__.py
@@ -1,10 +1,9 @@
-from .dynamic_biosphere_builder import DynamicBiosphereBuilder
 from .dynamic_characterization import DynamicCharacterization
-from .dynamic_biosphere_builder_from_timeline import DynamicBiosphereBuilder
+from .dynamic_biosphere_builder import DynamicBiosphereBuilder
 from .edge_extractor import EdgeExtractor
 from .matrix_modifier import MatrixModifier
 from .timeline_builder import TimelineBuilder
 from .timex_lca import TimexLCA
-from .SetList import SetList
+from .helper_classes import SetList
 
 __version__ = "0.1.4"
